ET use compiler macros in ops
Use new compiler macros to fix operator library builds.
Additional issues,
isnan and signbit supported types are only on float, double and long double. Apply static cast to avoid msvc compiler error.

Differential Revision: [D78854136](https://our.internmc.facebook.com/intern/diff/D78854136/)

ghstack-source-id: 305107633
Pull Request resolved: https://github.com/pytorch/executorch/pull/13638

+++ kernels/portable/cpu/op_sign.cpp
@@ -42,7 +42,7 @@ Tensor& sign_out(KernelRuntimeContext& ctx, const Tensor& in, Tensor& out) {
     ET_SWITCH_REALHBF16_TYPES(in.scalar_type(), ctx, "sign.out", CTYPE, [&] {
       apply_unary_map_fn(
           [](const CTYPE val_in) {
-            if (std::isnan(val_in)) {
+            if (std::isnan(static_cast<float>(val_in))) {
               return val_in;
             } else {
               return static_cast<CTYPE>((val_in > 0) - (val_in < 0));
